+++ src/fcitx-public.h
@@ -12,15 +12,6 @@ void start_fcitx_thread() noexcept;
 void stop_fcitx_thread() noexcept;
 void restart_fcitx_thread() noexcept;
 
-// Though being UInt, 32b is enough for modifiers
-bool process_key(ICUUID uuid, uint32_t unicode, uint32_t osxModifiers,
-                 uint16_t osxKeycode, bool isRelease) noexcept;
-
-ICUUID create_input_context(const char *appId) noexcept;
-void destroy_input_context(ICUUID uuid) noexcept;
-void focus_in(ICUUID uuid) noexcept;
-void focus_out(ICUUID uuid) noexcept;
-
 // NOTE: It's impossible to use std::vector<std::string> directly
 // until Swift fixes C++ interop.
 std::string input_method_groups() noexcept;
