+++ src/fcitx.h
@@ -7,7 +7,6 @@
 
 #include "fcitx-public.h"
 #include "../macosfrontend/macosfrontend.h"
-#include "webview_candidate_window.hpp"
 
 /// 'Fcitx' manages the lifecycle of the global Fcitx instance.
 class Fcitx final {
@@ -27,19 +26,14 @@ public:
     fcitx::Instance *instance();
     fcitx::AddonManager &addonMgr();
     fcitx::AddonInstance *addon(const std::string &name);
-    fcitx::MacosFrontend *macosfrontend();
 
 private:
     void setupLog(bool verbose);
     void setupEnv();
-    void setupCandidateWindow();
     void setupInstance();
-    void setupFrontend();
 
     std::unique_ptr<fcitx::Instance> instance_;
     std::unique_ptr<fcitx::EventDispatcher> dispatcher_;
-    fcitx::MacosFrontend *macosfrontend_;
-    std::unique_ptr<candidate_window::CandidateWindow> window_;
 };
 
 /// Run a function in the fcitx thread and obtain its return value
